+++ database/migrations/2024_05_01_010914_create_costos_variables_anuales_pesimistas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('Id_estudio_financiero')->constrained('estudio_financieros','id','idx_costos_variables_pesimista_estudio_financiero')->onDelete('cascade');
             $table->foreignId('Id_costo_variable')->constrained('costos_variables')->onDelete('cascade');
             $table->integer('mes');
-            $table->float('monto_pesimista');
+            $table->decimal('monto_pesimista',12,2);
             $table->timestamps();
         });
     }
