+++ database/migrations/2024_05_01_012312_create_ingresos_anuales_pesimistas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('Id_estudio_financiero')->constrained('estudio_financieros','id','idx_ingreso_pesimista_estudio_financiero')->onDelete('cascade');
             $table->foreignId('Id_ingresos')->constrained('ingresos')->onDelete('cascade');
             $table->integer('mes');
-            $table->float('monto_pesimista');
+            $table->decimal('monto_pesimista',12,2);
             $table->timestamps();
         });
     }
